+++ backend/src/task/entities/task.entity.ts
@@ -55,7 +55,7 @@ export class Task {
   // eslint-disable-next-line no-use-before-define
   dependent: Task[];
 
-  @ManyToMany(() => Task, task => task.dependent, { eager: true })
+  @ManyToMany(() => Task, task => task.dependent)
   // eslint-disable-next-line no-use-before-define
   dependencies: Task[];
 }
