+++ frontend/src/components/task/TaskList/index.tsx
@@ -1,18 +1,18 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import { Box, Grid, Typography } from '@mui/material';
-import TaskItem from '../TaskItem';
+import TaskItem from 'src/components/task/TaskItem';
 import { Task } from 'src/models/task.model';
 import { ModalHandler } from 'src/hooks/use-modals';
 import { CreateTaskDto } from 'src/services/dto/task/create-task.dto';
 import { toast } from 'react-toastify';
-import EditTaskForm from 'src/components/task/components/EditTaskForm';
+import EditTaskForm from 'src/components/task/EditTaskForm';
 import ModalForm from 'src/components/modalForm';
-import CreateTaskForm from 'src/components/task/components/CreateTaskForm';
+import CreateTaskForm from 'src/components/task/CreateTaskForm';
 import { UpdateTaskDto } from 'src/services/dto/task/update-task.dto';
 import { useStores } from 'src/hooks';
-import { styles } from 'src/components/task/components/TaskList/styles';
-import TaskSort from 'src/components/task/components/TaskSort';
-import DeleteTaskForm from 'src/components/task/components/DeleteTaskModal';
+import { styles } from 'src/components/task/TaskList/styles';
+import TaskSort from 'src/components/task/TaskSort';
+import DeleteTaskForm from 'src/components/task/DeleteTaskModal';
 import { observer } from 'mobx-react';
 import { Project } from 'src/models/project.model';
 import { UserStore } from 'src/stores/user.store';
